função newGame passada para Game.js
